Changed graph readings to time vs RSSI Completed Measure Threshold speed functionality

+++ src/app/AutoSense/app/src/main/java/org/autosense/commons/AppConfig.java
@@ -31,7 +31,7 @@ public class AppConfig {
         entries.clear();
     }
 
-    public void setEntry(int i, int rssi) {
+    public void setEntry(float i, int rssi) {
         entries.add(new Entry(i, rssi));
     }
